chore: migrated loader to packages
