+++ packages/components/src/Loader/Loader.stories.tsx
@@ -1,0 +1,9 @@
+import { Loader } from './Loader'
+import type { ComponentStory, ComponentMeta } from '@storybook/react'
+
+export default {
+  title: 'Base Components/Loader',
+  component: Loader,
+} as ComponentMeta<typeof Loader>
+
+export const Default: ComponentStory<typeof Loader> = () => <Loader />
