+++ packages/components/src/index.ts
@@ -19,3 +19,4 @@ export { NotificationItem } from './NotificationItem/NotificationItem'
 export type { UserNotificationList } from './NotificationList/NotificationList'
 export { NotificationList } from './NotificationList/NotificationList'
 export { CommentList } from './CommentList/CommentList'
+export { Loader } from './Loader/Loader'
