--- src/components/Loader/index.tsx
@@ -1,61 +1,0 @@
-import { Component } from 'react'
-import { Flex, Image, Text } from 'theme-ui'
-import styled from '@emotion/styled'
-import { keyframes } from '@emotion/react'
-
-import { inject, observer } from 'mobx-react'
-import type { ThemeStore } from 'src/stores/Theme/theme.store'
-
-interface IProps {
-  isMobile?: boolean
-}
-
-const rotate = keyframes`
-  from {
-    transform: rotate(0deg);
-  }
-
-  to {
-    transform: rotate(360deg);
-  }
-`
-
-const RotatingLogo = styled(Image)`
-  animation: ${rotate} 2s cubic-bezier(0.445, 0.05, 0.55, 0.95) infinite;
-  padding: 1rem;
-`
-
-@inject('themeStore')
-@observer
-export class Loader extends Component<IProps> {
-  // eslint-disable-next-line
-  constructor(props: any) {
-    super(props)
-  }
-
-  get injected() {
-    return this.props as {
-      themeStore: ThemeStore
-    }
-  }
-
-  render() {
-    const logo = this.injected.themeStore?.currentTheme?.logo || null
-    return (
-      <>
-        <Flex sx={{ flexWrap: 'wrap', justifyContent: 'center' }}>
-          {logo && (
-            <RotatingLogo
-              loading="lazy"
-              src={logo}
-              sx={{ width: [75, 75, 100] }}
-            />
-          )}
-          <Text sx={{ width: '100%', textAlign: 'center' }}>loading...</Text>
-        </Flex>
-      </>
-    )
-  }
-}
-
-export default Loader
