+++ src/pages/Howto/Content/HowtoList/HowtoList.tsx
@@ -1,9 +1,8 @@
 import { inject, observer } from 'mobx-react'
 import * as React from 'react'
 import { AuthWrapper } from 'src/components/Auth/AuthWrapper'
-import { Button, MoreContainer } from 'oa-components'
+import { Button, MoreContainer, Loader } from 'oa-components'
 import { Heading, Input, Flex, Box } from 'theme-ui'
-import { Loader } from 'src/components/Loader'
 import TagsSelect from 'src/components/Tags/TagsSelect'
 import { VirtualizedFlex } from 'src/pages/Howto/VirtualizedFlex/VirtualizedFlex'
 import type { IHowtoDB } from 'src/models/howto.models'
