+++ src/pages/Research/Content/CreateUpdate/index.tsx
@@ -2,7 +2,7 @@ import { observer } from 'mobx-react'
 import * as React from 'react'
 import type { RouteComponentProps } from 'react-router'
 import { Redirect } from 'react-router'
-import Loader from 'src/components/Loader'
+import { Loader } from 'oa-components'
 import { Text } from 'theme-ui'
 import { useResearchStore } from 'src/stores/Research/research.store'
 import { isAllowToEditContent } from 'src/utils/helpers'
