+++ src/pages/Research/Content/EditResearch/index.tsx
@@ -3,7 +3,7 @@ import { observer } from 'mobx-react'
 import * as React from 'react'
 import type { RouteComponentProps } from 'react-router'
 import { Redirect } from 'react-router'
-import { Loader } from 'src/components/Loader'
+import { Loader } from 'oa-components'
 import { Text } from 'theme-ui'
 import type { IResearch } from 'src/models/research.models'
 import type { IUser } from 'src/models/user.models'
