+++ src/pages/Research/Content/ResearchArticle.tsx
@@ -2,8 +2,7 @@ import { observer } from 'mobx-react'
 import * as React from 'react'
 import type { RouteComponentProps } from 'react-router'
 import { Box, Flex } from 'theme-ui'
-import { Button } from 'oa-components'
-import { Loader } from 'src/components/Loader'
+import { Button, Loader } from 'oa-components'
 import { NotFoundPage } from 'src/pages/NotFound/NotFound'
 import { useResearchStore } from 'src/stores/Research/research.store'
 import { isAllowToEditContent } from 'src/utils/helpers'
